use a TreeGrower to spawn the tea tree
Also spawn a tea tree more often
