+++ src/main/java/knightminer/simplytea/block/TeaTreeGrower.java
@@ -1,0 +1,16 @@
+package knightminer.simplytea.block;
+
+import knightminer.simplytea.core.Registration;
+import net.minecraft.core.Holder;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.block.grower.AbstractTreeGrower;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+
+public class TeaTreeGrower extends AbstractTreeGrower {
+
+	@Override
+	protected Holder<? extends ConfiguredFeature<?, ?>> getConfiguredFeature(RandomSource pRandom, boolean pLargeHive) {
+		return Registration.configured_tea_tree.getHolder().get();
+	}
+
+}
